+++ ReleaseNotes.md
@@ -8,6 +8,7 @@
 - cta/CTA#999 - Add a default mount rule for recalls
 - cta/CTA#1109 - Add --dirtybit option to cta-admin ta ch and show dirty bit value in cta-admin --json ta ls
 - cta/CTA#1111 - Add disk instance and disk instance tables to catalogue and respective cta-admin diskinstance/diskinstancespace add/ls/ch/rm commands
+- cta/CTA#1108 - Make cta-admin --json sq show mount policy with highest priority and mount policy with lowest request age for each queue.
 
 ## Bug fixes
 - cta/CTA#1102 - Make requeued jobs retain their original creation time